Remove use of sed in GIT_EDITOR
It should not be needed as the comment lines will be removed.

The previous version also conflicts with Git option `commit.verbose =
true` which adds the diff below the line:

  # Do not modify or remove the line above.
  # Everything below it will be ignored.

and since the diff does not begin with `#` it gets added to the squashed
commit which is not what we want.